The optical flow is sort of working, though the output image is scaled down with a screencap apparently running in half of it (or just other memory).  Probably an 8UC4 thing?
Comments prior to making these changes:

Make a cv::farneback optical flow node with two image inputs (also possible an input for the initial optical flow, but the differing datatype is a problem- but since the flow input is just an initial guess then loss of precision from converting back and forth to CV8UC may not be that big of deal), and have interp signal input that will generate an intermediate image (also passing in > 1.0 or < 0.0 will make an image projected into the future).  It would be nice if the remap could accept the optical flow map and generate the interp image but that might be too difficult.  Also all the other parameters would be available, since the point is to be able to explore the parameter space and find good slow motion generation inputs.