OpenGL node, doesn't interact with other nodes yet. TBD import and export texture and provide signals to control vertices.
The opengl node can never be truly general purpose without a lot of work to support vectors of points and indices being passed around in signal buffers, which is probably a worthy goal, for the near future it will support basic primitives for display as surrounding background objects and foreground objects and a few texture image inputs and a single glReadPixel image output.  It may be put into a dedicated thread for performance.