Simplex noise http://code.google.com/p/battlestar-tux/source/browse/procedural/simplextextures.cpp
TBD Need to save more state with save, and autoincrement filenames.  A symlink can be maintained that points to the most recent so it can be loaded automatically, or search through the directory for the latest filename.

The entirety of buffers need to be saved to disk and reloaded.  Every save will create a new directory with one layer beneath with image buffer images (everything else could be at the top level, 1 image or 1 signal buffer yaml file per node.

TBD normalize all node inputs that are related resolution - Rot2D especially