Not yet compiling.
Adding string as fundamental type, which means arrays of strings also need to be included.  The proliferation of similar code really begs for a better general implementation.

This is also the beginning of adding a file selector to the ImageDir or a node that feeds it the string location of the directory.